requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.6
click==8.0.1
cmake==3.21.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-environ==0.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
face-recognition-models==0.3.0
Flask==2.0.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.3
opencv-contrib-python==4.5.4.58
passlib==1.7.4
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pytz==2021.1
razorpay==1.2.0
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
twilio==7.0.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
zope.interface==5.4.0
